requirements.txt
# Daily Briefing - Dependencies

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# RSS Parsing
feedparser>=6.0.10

# Web Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# LLM & AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.12.0
anthropic>=0.18.0

# Embeddings & Vector
sentence-transformers>=2.3.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Data Processing
pyyaml>=6.0.1
python-dateutil>=2.8.2
pytz>=2024.1

# Database
aiosqlite>=0.19.0

# Image Processing
Pillow>=10.2.0

# Feishu SDK
lark-oapi>=1.0.0

# Markdown
markdown>=3.5.0

# Utilities
tenacity>=8.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Translation (backup)
deep-translator>=1.11.0

# Content Extraction
trafilatura>=1.6.0
readability-lxml>=0.8.1
